% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1normFun.R
\name{L1normFun}
\alias{L1normFun}
\title{L1normFun}
\usage{
L1normFun(x)
}
\arguments{
\item{x}{vector}
}
\value{
The \eqn{l_1} norm of vector x
}
\description{
The function aims to compute the \eqn{l_1} norm.
}
\details{
The L1normFun aims to compute the \eqn{l_1} norm: \eqn{\sum_i^n |x_i|}
}
\author{
Xinlin Hu \href{mailto:thompson-xinlin.hu@connect.polyu.hk}{thompson-xinlin.hu@connect.polyu.hk}

Yaohua Hu \href{mailto:mayhhu@szu.edu.cn}{mayhhu@szu.edu.cn}
}
